requirements.txt
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
ccxt==4.4.86
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
diff-match-patch==20241021
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-import-export==4.3.7
djangorestframework==3.16.0
djangorestframework-api-key==3.1.0
drf-spectacular==0.28.0
frozenlist==1.6.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
multidict==6.4.4
packaging==25.0
propcache==0.3.1
pycares==4.8.0
pycparser==2.22
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
setuptools==80.9.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
yarl==1.20.0
